requirements.txt
asttokens==2.4.1
backcall==0.2.0
blosc==1.11.1
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.7
executing==2.0.1
fonttools==4.43.1
importlib-resources==6.1.0
impyute==0.0.8
ipython==8.12.3
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
lightgbm==4.1.0
matplotlib==3.7.3
matplotlib-inline==0.1.6
miceforest==5.6.3
numpy==1.24.4
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.2.0
traitlets==5.12.0
typing-extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.8
yattag==1.15.2
zipp==3.17.0
